wip: scream uploader
Adds a component and the necessary firebase things to allow scream
uploads. The admin fills in scream name and note info into the form. The
info is saved in Firebase Firestore. The file is then uploaded to
Firebase Storage with the name being the random id generated for the
Firestore document. Upon upload, the public url to the file is saved to
the Firestore document.

WIP will change later.